tools/python/airmaps/requirements.txt
apache-airflow [postgres]==1.10.10
psycopg2-binary==2.8.4
cryptography==2.8
webdavclient==1.0.8
-r ../maps_generator/requirements.txt